Add --display option
gpg-agent might call pinentry-program with additional argument:

    --display /private/tmp/com.apple.launchd.weZyuRW3RM/org.macosforge.xquartz:0

Since bitwarden-pinentry does not recognized and parse this
--display argument, the program will panic due to the unknown
argument error.

This approach adds the --display argument into bitwarden-pinentry
to mitigate the panic but ignore the value of that argument.

go-arg is chosen as the argument/flag parser instead of the Go's
stdlib Flag package to handle the double dashes (--) arguments
implementation as oppose to the default single dash. This also
causes breaking change as the config argument has to be called as
--config instead of -config.